requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
qdrant-client==1.17.1
requests==2.33.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyMuPDF==1.27.2.2
google-genai==1.70.0